refactor: move all media related api calls to api folder
